requirements.txt
Flask>=0.10.1
Flask-Bootstrap>=3.3.0.1
Flask-WTF>=0.11
futures>=2.2.0
paho-mqtt>=1.1
pyownet>=0.8.1
PyDispatcher>=2.0.5
WTForms-Alchemy>=0.13.0
blinker>=0.0